+++ template/server/server.ts
@@ -5,24 +5,13 @@ import { setupSampleLakebaseRoutes } from './routes/lakebase/todo-routes';
 
 createApp({
   plugins: [
-{{- if .plugins.lakebase}}
-    server({ autoStart: false }),
-{{- range $name, $_ := .plugins}}
-{{- if ne $name "server"}}
-    {{$name}}(),
-{{- end}}
-{{- end}}
-{{- else}}
 {{- range $name, $_ := .plugins}}
     {{$name}}(),
-{{- end}}
 {{- end}}
   ],
-})
 {{- if .plugins.lakebase}}
-  .then(async (appkit) => {
+  async onPluginsReady(appkit) {
     await setupSampleLakebaseRoutes(appkit);
-    await appkit.server.start();
-  })
+  },
 {{- end}}
-  .catch(console.error);
+}).catch(console.error);